requirements_dev.txt
pip==21.1
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
flake8==3.7.8
tox==3.14.0
coverage==4.5.4
sphinx<5,>=3.1
twine==1.14.0
Click==7.0
pytest==6.2.5
pytest-runner==5.1
mypy==0.910
jupyterlab==3.2.4
m2r2==0.3.2
duckdb-engine==0.1.8
testing.postgresql==1.3.0
psycopg2-binary==2.9.3
